[refactor] DateUtil 합치기 (#55)
