+++ src/main/java/konkuk/thip/common/util/DateUtil.java
@@ -3,6 +3,7 @@ package konkuk.thip.common.util;
 import org.springframework.stereotype.Component;
 
 import java.time.Duration;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Component
@@ -16,6 +17,29 @@ public class DateUtil {
         long hours = minutes / 60;
         if (hours < 24) return hours + "시간 전";
         return (hours / 24) + "일 전";
+    }
+
+    public static String formatAfterTime(LocalDate date) {
+        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime dateTime = date.atStartOfDay();
+        Duration d = Duration.between(now, dateTime);
+
+        if (d.isNegative() || d.isZero()) {
+            return "??";
+        }
+
+        long days = d.toDays();
+        if (days > 0) {
+            return days + "일 뒤 ";
+        }
+
+        long hours = d.toHours();
+        if (hours > 0) {
+            return hours + "시간 뒤 ";
+        }
+
+        long minutes = d.toMinutes();
+        return minutes + "분 뒤 ";
     }
 }
